# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Nulls Report production portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      tags: [account]
      summary: Get the signed-in portal user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateCurrentUser
      tags: [account]
      summary: Update the signed-in portal user's account details
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /admin/users:
    get:
      operationId: listPortalUsers
      tags: [administration]
      summary: List portal users for administrators
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      operationId: clearPortalUsers
      tags: [administration]
      summary: Clear the entire user database (all accounts, reports, messages, notifications)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClearDatabaseInput"
      responses:
        "200":
          description: Database cleared
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}/block:
    patch:
      operationId: updatePortalUserBlock
      tags: [administration]
      summary: Block or unblock a portal user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUpdateInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/users/{id}:
    delete:
      operationId: deletePortalUser
      tags: [administration]
      summary: Remove a user and all their data from the portal
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed: { type: number }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/users/{id}/role:
    patch:
      operationId: updatePortalUserRole
      tags: [administration]
      summary: Update a portal user role
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdateInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: number
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Insufficient permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id: { type: number }
        tag: { type: ["string", "null"] }
        displayName: { type: string }
        role:
          type: string
          enum: [user, moderator, administrator]
        blocked: { type: boolean }
        nullsConnectId: { type: ["string", "null"] }
        nullsConnectName: { type: ["string", "null"] }
        avatarPath: { type: ["string", "null"] }
        discordId: { type: ["string", "null"] }
        discordUsername: { type: ["string", "null"] }
        preferences:
          type: object
          additionalProperties: true
        authMethods:
          type: array
          items: { $ref: "#/components/schemas/AuthMethod" }
        hasTrustedAuth: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, tag, displayName, role, preferences, authMethods, hasTrustedAuth, createdAt]
    AuthMethod:
      type: object
      properties:
        provider: { type: string, enum: [discord, nulls_connect] }
        name: { type: string }
        label: { type: ["string", "null"] }
        linkedAt: { type: string, format: date-time }
      required: [provider, name, label, linkedAt]
...
    RoleUpdateInput:
      type: object
      properties:
        role:
          type: string
          enum: [user, moderator, administrator]
      required: [role]
    BlockUpdateInput:
      type: object
      properties:
        blocked: { type: boolean }
      required: [blocked]
    ClearDatabaseInput:
      type: object
      properties:
        confirm: { type: boolean }
      required: [confirm]
...
    UpdateUserInput:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 80 }
        preferences:
          type: object
          additionalProperties: true
      description: At least one field is required.